requirements.txt
# Core dependencies
anthropic==0.39.0
openai==1.54.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

# Database
sqlalchemy==2.0.36
sqlite3-python==1.0.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Image processing
Pillow==11.0.0
matplotlib==3.10.0
plotly==5.24.1

# API clients
wordpress-api==1.2.9
feedparser==6.0.11

# Scheduling
schedule==1.2.2
apscheduler==3.10.4

# Utils
python-dateutil==2.9.0
pytz==2024.2
colorama==0.4.6
tqdm==4.67.1

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0

# Logging
loguru==0.7.3

# Data validation
pydantic==2.10.3
jsonschema==4.23.0

# Web scraping
selenium==4.27.1
undetected-chromedriver==3.6.1

# SEO
yake==0.4.8
textstat==0.7.4